+++ src/analyze/src/components/utils/utils.js
@@ -1,5 +1,0 @@
-export default function ensureObjPropIsReactive(vueSetFn, obj, prop, val = false) {
-  if (!obj.prop) {
-    vueSetFn(obj, prop, val);
-  }
-}